+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 Release Date: _unreleased_
 
 - Changed default configuration value of MaxPageSize for endpoints to 100. Validate if max page size is higher than DefaultPageSize.
-- Added `/transaction/outcomes` endpoint - a proxy to Node's Core API # TODO TODO document
+- Added `/transaction/committed-outcomes` endpoint returning fungible and non-fungible balance changes for a given transaction.
 - Added new opt-in `receipt_output` to `/stream/transactions`, and `/transaction/committed-details` endpoints. Temporarily set by default to true, to allow client's migration.
 - Optimized `TransactionQuerier.GetTransactions` not to fetch unnecessary data from underlying database.